pyproject.toml
[project]
name = "x-automation-studio"
version = "0.1.0"
description = "A platform for LLM-assisted social media marketing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests-oauthlib>=1.3.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
    "fastapi>=0.115.6",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.5",
    "mypy>=1.14.1",
    "types-requests>=2.32.0.20241016",
    "wonderwords>=2.2.0",
    "litellm>=1.59.9",
]
